Move RuleLog and related types to own module
